uc: Update tf_hal_log_message prototypes.

+++ uc/hal_raspberry_pi/hal_raspberry_pi.c
@@ -94,7 +94,7 @@ TF_HalCommon *tf_hal_get_common(TF_HalContext *hal) {
     return &hal->hal_common;
 }
 
-void tf_hal_log_message(const char *msg, uint32_t len) {
+void tf_hal_log_message(const char *msg, size_t len) {
     fwrite(msg, len, 1, stdout);
 }
